pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.10"
name = "topolink"
version = "0.1.0"
description = "A Python package for deploying distributed undirected network topologies on a set of nodes."
authors = [{ name = "Rui Huang", email = "R.Huang@lboro.ac.uk" }]

dependencies = ["numpy>=2.2.6", "pyzmq>=27.0.0"]

[tool.setuptools.packages.find]
where = ["src"]
